requirements.txt
alembic==1.4.2
attrs==19.3.0
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
wcwidth==0.2.4
Werkzeug==1.0.1
